- Governance topics now support using Discord's reply feature to reply to a given thread message. The reply will indent the original message and author's username similar to other forums, and place the new reply underneath. - Added new command /tournamentstatus which shows the current tournament week's status. This includes the stage its on (i.e. sign-ups), who is participating, the achievement to complete, the rewards for that week, any sponsors, and the current winners - Notification added to #feedback channel with all available governance topics to discuss and vote on. - Faucet is now one generic 24 hour period (based on UTC) rather than the original 24 hour timer per user - Added fix to stop thread messages from being added if someone typed in a governance channel without being registered. - Command cooldowns now show time remaining between using each command similar to how faucet worked before - Added fix for Discord channel ordering to properly shift a recently updated governance thread to the top of the category - Added message when someone votes on a governance topic - Fully added rewards for any governance topic that survives the full 30 day window (without being deleted due to inactivity over 7 days) and has a majority of yes votes - Updated most bot posts or responses as embeds to help support readability - Added additional logging to a few areas to help support future debugging - Added Discord markup around links in random bot messaging to stop embeds from constantly showing up - Added new commands to /help command list

+++ EMBRS/Database/DatabaseThreads.cs
@@ -55,7 +55,7 @@ namespace EMBRS
                 .WithAuthor(author.ToString(), author.GetAvatarUrl() ?? author.GetDefaultAvatarUrl())
                 .WithTitle(linkedThread.GetThreadHeader())
                 .WithDescription(linkedThread.GetThreadContent())
-                .AddField("Yes Votes", "0")
+                .AddField("Yes Votes", "0", true)
                 .AddField("No Votes", "0", true)
                 .WithCurrentTimestamp()
                 .WithColor(Color.Orange);
@@ -67,21 +67,40 @@ namespace EMBRS
         public async Task UpdateThreadPositionInChannel(Thread thread, DiscordSocketClient client = null)
         {
             var guild = client.GetGuild(ulong.Parse(Settings.GuildID));
-            var threadChannel = guild.TextChannels.FirstOrDefault(x => x.Id == thread.GetThreadChannelId());
-            await threadChannel.ModifyAsync(x =>
+            var category = guild.CategoryChannels.FirstOrDefault(category => category.Id == _categoryId);
+            var categoryChannels = category.Channels.ToList();
+            var indexOfThread = 0;
+
+            for (int i = 0; i < categoryChannels.Count; i++)
             {
-                x.Position = 0;
-            });
+                if(categoryChannels[i].Id == thread.GetThreadChannelId())
+                {
+                    indexOfThread = i;
+                    break;
+                }
+            }
+
+            var threadChannel = categoryChannels[indexOfThread];
+            categoryChannels.RemoveAt(indexOfThread);
+            categoryChannels.Insert(0, threadChannel);
+
+            for (int i = 0; i < categoryChannels.Count; i++)
+            {
+                await categoryChannels[i].ModifyAsync(x =>
+                {
+                    x.Position = i;
+                });
+            }
         }
 
-        public async Task<bool> ContainsThreadByChannelId(ulong channelId)
+        public bool ContainsThreadByChannelId(ulong channelId)
         {
             return _registeredThreads.Any(thread => thread.Value.GetThreadChannelId() == channelId);
         }
 
         public async Task<Thread> GetThreadByChannelId(ulong channelId)
         {
-            if (await ContainsThreadByChannelId(channelId)) return _registeredThreads.FirstOrDefault(thread => thread.Value.GetThreadChannelId() == channelId).Value;
+            if (ContainsThreadByChannelId(channelId)) return _registeredThreads.FirstOrDefault(thread => thread.Value.GetThreadChannelId() == channelId).Value;
             return null;
         }
 
@@ -91,22 +110,93 @@ namespace EMBRS
             if(thread != null && _registeredThreads.ContainsKey(thread.GetThreadId())) _registeredThreads.Remove(thread.GetThreadId());
         }
 
+        public List<Thread> GetAllThreads()
+        {
+            var threads = new List<Thread>();
+            foreach (var thread in _registeredThreads) threads.Add(thread.Value);
+            return threads;
+        }
+
         public async Task TestAllThreads(DiscordSocketClient client)
         {
             var threadsToDelete = new List<Thread>();
-            foreach(var thread in _registeredThreads)
+            foreach (var thread in _registeredThreads)
             {
-                if((DateTime.UtcNow - thread.Value.GetThreadUpdateTime()).TotalDays >= Settings.IdleThreadTimeToDelete && thread.Value.GetThreadExpirationTime() > DateTime.UtcNow)
+                // HANDLE DELETE
+                if (((DateTime.UtcNow - thread.Value.GetThreadUpdateTime()).TotalDays >= Settings.IdleThreadTimeToDelete && thread.Value.GetThreadExpirationTime() > DateTime.UtcNow) ||
+                    thread.Value.GetThreadChannelId() == 0)
                 {
                     threadsToDelete.Add(thread.Value);
                 }
+
+                // HANDLE REWARD
+                if (DateTime.UtcNow > thread.Value.GetThreadExpirationTime() && !thread.Value.GetThreadRewarded())
+                {
+                    var yesVotes = 0;
+                    var noVotes = 0;
+
+                    var author = thread.Value.GetThreadAuthor();
+
+                    var rewardedUsers = new List<ulong>();
+                    foreach (var vote in thread.Value.GetVotes())
+                    {
+                        if (vote.Value) yesVotes++;
+                        else noVotes++;
+                        if(vote.Key != author) rewardedUsers.Add(vote.Key);
+                    }
+
+                    foreach (var threadMessage in thread.Value.GetThreadDiscussion())
+                    {
+                        if (!rewardedUsers.Contains(threadMessage.GetThreadMessageAuthor()) && threadMessage.GetThreadMessageAuthor() != author) rewardedUsers.Add(threadMessage.GetThreadMessageAuthor());
+                    }
+
+                    if(yesVotes > noVotes)
+                    {
+                        thread.Value.SetThreadRewarded(true);
+                        var guild = client.GetGuild(ulong.Parse(Settings.GuildID));
+                        var authorUser = guild.GetUser(author);
+                        var embrsChannel = guild.TextChannels.FirstOrDefault(x => x.Name == "embrs");
+
+                        await XRPL.SendRewardAsync(client, null, null, authorUser, "100");
+                        Database.GetDatabase<DatabaseAccounts>(DatabaseType.Accounts).GetAccount(authorUser.Id).ModEMBRSEarned(100);
+
+                        for (int i = 0; i < rewardedUsers.Count; i++)
+                        {
+                            var user = guild.GetUser(rewardedUsers[i]);
+                            await XRPL.SendRewardAsync(client, null, null, user, "10");
+                            Database.GetDatabase<DatabaseAccounts>(DatabaseType.Accounts).GetAccount(user.Id).ModEMBRSEarned(10);
+                        }
+
+                        var embedBuiler = new EmbedBuilder()
+                            .WithAuthor(authorUser.ToString(), authorUser.GetAvatarUrl() ?? authorUser.GetDefaultAvatarUrl())
+                            .WithTitle(thread.Value.GetThreadHeader())
+                            .WithDescription(thread.Value.GetThreadContent())
+                            .AddField("Successful Governance Topic", "The above topic has made it through 30 days of discussion and has a majority of yes votes! Rewards have been sent out to the topic author and anyone who discussed or voted in it.")
+                            .WithCurrentTimestamp()
+                            .WithColor(Color.Orange);
+
+                        var message = await embrsChannel.SendMessageAsync(null, false, embedBuiler.Build());
+                    }
+                }
             }
 
-            for(int i = 0; i < threadsToDelete.Count; i++)
+            for (int i = 0; i < threadsToDelete.Count; i++)
             {
-                var guild = client.GetGuild(ulong.Parse(Settings.GuildID));
-                var channel = guild.TextChannels.FirstOrDefault(x => x.Id == threadsToDelete[i].GetThreadChannelId());
-                await channel.DeleteAsync();
+                if (threadsToDelete[i].GetThreadChannelId() != 0)
+                {
+                    var guild = client.GetGuild(ulong.Parse(Settings.GuildID));
+                    var channel = guild.TextChannels.FirstOrDefault(x => x.Id == threadsToDelete[i].GetThreadChannelId());
+                    await channel.DeleteAsync();
+                }
+                else
+                {
+                    if (_registeredThreads.ContainsKey(threadsToDelete[i].GetThreadId())) _registeredThreads.Remove(threadsToDelete[i].GetThreadId());
+                }
+            }
+
+            foreach (var thread in _registeredThreads)
+            {
+
             }
         }
     }
